+++ src/services/shop/models.rs
@@ -2,6 +2,6 @@ pub struct ShopItem {
     pub id: i32,
     pub name: &'static str,
     pub description: &'static str,
-    pub price: f64,
+    pub price: i64,
     pub role_id: Option<u64>,
 }
